ReSwiTS/requirements.txt
yfinance
pandas
numpy
scipy
matplotlib
plotly
scikit-learn
hmmlearn
ta
pyyaml
streamlit